price data type change

+++ src/main/java/com/practice/dto/BookDto.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -19,6 +21,6 @@ public class BookDto {
 
     private String bookAuthor;
 
-    private Integer bookPrice;
+    private BigDecimal bookPrice;
 
 }
